+++ database/migrations/2020_12_11_022054_create_reviews_table.php
@@ -15,14 +15,15 @@ class CreateReviewsTable extends Migration
     {
         Schema::create('reviews', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('user_id')->unsigned()->index();
-            $table->integer('book_id')->unsigned()->index();
+            $table->Integer('user_id')->unsigned()->index();
+            $table->bigInteger('book_id')->unsigned()->index();
             $table->text('body');
             $table->integer('is_active')->default(0);
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->foreign('book_id')->references('id')->on('books')->onDelete('cascade');
+
         });
     }
